Ensure that pool contain enough problems
If a pool is used n times within a test-setup we need at
least n problems so no problem is repeated within a group.

Before this commit this was only ensured within a single
test. Now, it is ensured within all tests of a group. If
the condition is not fulfilled an Exception is raised.