Moved seed value in seetings file
It must be possible to generate the same exams from run to run.
Setting the seed value from the command line is error-prone. Hence,
this option has been moved into the settings file. If no seed value
is specified in the settings file the seed is generated from the
system date.